Split the Politeia and agendas further
