+++ gov/politeia/log.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2018, The Decred developers
 // See LICENSE for details.
 
-package offchaindb
+package politeia
 
 import "github.com/decred/slog"
