+++ views/proposals.tmpl
@@ -45,10 +45,10 @@
                         {{end}}
                     </td>
                     <td>
-                        {{if ne .PubishedDate 0}}
-                            {{TimeConversion .PubishedDate}}
+                        {{if ne .PublishedDate 0}}
+                            {{TimeConversion .PublishedDate}}
                         {{else}}
-                            {{.PubishedDate}}
+                            {{.PublishedDate}}
                         {{end}}
                     </td>
                 </tbody>
